+++ changelog.md
@@ -1,3 +1,12 @@
+## 0.39 Alpha
+
+- Slightly better error message when multiselecting
+- Some render optimizations
+- Add button to return to main page after game over or not logged in
+- Bug fix: don't allow undo of automatic reaction fire pass
+- Bug fix: fixed bug with calculated average rating
+- Updated 308
+
 ## 0.38 Alpha
 
 - Added high contrast palette for colorblindness